Moved the layout to the page component
