+++ modules/app/recording/dashboard.tsx
@@ -3,34 +3,22 @@ import React, { useState, useRef, useEffect, useCallback } from "react";
 import { useRouter } from "next/router";
 import Image from "next/image";
 import { useTranslations } from "next-intl";
-import Navbar from "@/resources/containers/nav";
-import SideBar from "@/resources/containers/sidebar";
-import dash_styles from "@/styles/pages/dashboard.module.css";
 import { getStaticPropsWithTranslations } from "@/modules/lang/props";
 import messageHandler from "@/core/message-handler";
-import { AudioRecorder } from "@/core/audio-recorder";
+import { AudioRecorder } from "@/modules/app/recording/libs/audio-recorder";
 import { GetStaticProps } from "next";
+import dash_styles from "@/styles/pages/dashboard.module.css";
 
 export const getStaticProps: GetStaticProps = getStaticPropsWithTranslations;
 
 const Dashboard = () => {
   const t = useTranslations("");
   const router = useRouter();
-  const [sidebarOpen, setSidebarOpen] = useState(false);
   const [isRecording, setIsRecording] = useState(false);
   const [isDragging, setIsDragging] = useState(false);
   const [selectedFile, setSelectedFile] = useState<File | null>(null);
   const audioRecorderRef = useRef<AudioRecorder | null>(null);
   const fileInputRef = useRef<HTMLInputElement>(null);
-
-  const sideBarItems = [
-    { icon: "/recordings.svg", label: "Recordings" },
-    { icon: "/history.svg", label: "Historical" },
-  ];
-
-  const toggleSidebar = () => {
-    setSidebarOpen(!sidebarOpen);
-  };
 
   const toggleRecording = async () => {
     if (!audioRecorderRef.current) {
@@ -130,100 +118,91 @@ const Dashboard = () => {
   };
 
   return (
-    <div className={dash_styles.container}>
-      <Navbar toggleSidebar={toggleSidebar} />
-      <div className={dash_styles.contentWrapper}>
-        <SideBar
-          _activeTab={"Recordings"}
-          isOpen={sidebarOpen}
-          closeSidebar={() => setSidebarOpen(false)}
-          sideBarItems={sideBarItems}
-        />
-        <main className={dash_styles.mainContent}>
-          <h2 className={dash_styles.h2}>
-            {t(
-              "Record your consultation or upload an audio with the previously recorded consultation to generate a report",
-            )}
-          </h2>
-          <p className={`${dash_styles.p} ${dash_styles.top_p}`}>
-            {t("Activate the audio recorder")}
+    <>
+      <h2 className={dash_styles.h2}>
+        {t(
+          "Record your consultation or upload an audio with the previously recorded consultation to generate a report"
+        )}
+      </h2>
+      <p className={`${dash_styles.p} ${dash_styles.top_p}`}>
+        {t("Activate the audio recorder")}
+      </p>
+
+      <div className={dash_styles.contentColumns}>
+        <div className={dash_styles.recordSection}>
+          <button
+            className={`${dash_styles.recordButton} ${
+              isRecording ? dash_styles.recording : ""
+            }`}
+            onClick={toggleRecording}
+          >
+            <Image
+              src="/recordings.svg"
+              alt="Microphone"
+              width={60}
+              height={60}
+            />
+          </button>
+          <p className={dash_styles.p}>
+            {isRecording
+              ? t("Click To Stop Recording")
+              : t("Click To Start Recording")}
           </p>
+        </div>
 
-          <div className={dash_styles.contentColumns}>
-            <div className={dash_styles.recordSection}>
-              <button
-                className={`${dash_styles.recordButton} ${isRecording ? dash_styles.recording : ""}`}
-                onClick={toggleRecording}
-              >
-                <Image
-                  src="/recordings.svg"
-                  alt="Microphone"
-                  width={60}
-                  height={60}
-                />
-              </button>
-              <p className={dash_styles.p}>
-                {isRecording
-                  ? t("Click To Stop Recording")
-                  : t("Click To Start Recording")}
-              </p>
-            </div>
+        <div className={dash_styles.divider}></div>
 
-            <div className={dash_styles.divider}></div>
-
-            <div className={dash_styles.uploadSection}>
-              <h3 className={dash_styles.h3}>{t("Upload")}</h3>
-              <div
-                className={`${dash_styles.uploadArea} ${isDragging ? dash_styles.dragging : ""}`}
-                onDragEnter={handleDragEnter}
-                onDragOver={handleDragOver}
-                onDragLeave={handleDragLeave}
-                onDrop={handleDrop}
+        <div className={dash_styles.uploadSection}>
+          <h3 className={dash_styles.h3}>{t("Upload")}</h3>
+          <div
+            className={`${dash_styles.uploadArea} ${
+              isDragging ? dash_styles.dragging : ""
+            }`}
+            onDragEnter={handleDragEnter}
+            onDragOver={handleDragOver}
+            onDragLeave={handleDragLeave}
+            onDrop={handleDrop}
+          >
+            <Image
+              src="/cloud-avatar.svg"
+              alt="Upload"
+              width={50}
+              height={50}
+            />
+            <p className={dash_styles.p}>
+              {t("Drag & Drop Or")}{" "}
+              <span
+                onClick={handleBrowseClick}
+                className={dash_styles.browseLink}
               >
-                <Image
-                  src="/cloud-avatar.svg"
-                  alt="Upload"
-                  width={50}
-                  height={50}
-                />
-                <p className={dash_styles.p}>
-                  {t("Drag & Drop Or")}{" "}
-                  <span
-                    onClick={handleBrowseClick}
-                    className={dash_styles.browseLink}
-                  >
-                    {t("Browse")}
-                  </span>
-                </p>
-                <small className={dash_styles.small}>
-                  {t("Supported Formats: MP3")}
-                </small>
-                {selectedFile && (
-                  <p className={dash_styles.selectedFile}>
-                    {selectedFile.name}
-                  </p>
-                )}
-              </div>
-              <input
-                type="file"
-                ref={fileInputRef}
-                onChange={handleFileInput}
-                accept="audio/*"
-                style={{ display: "none" }}
-              />
-              <button
-                className={dash_styles.uploadButton}
-                onClick={handleUpload}
-                disabled={!selectedFile}
-              >
-                {t("Upload Files")}
-              </button>
-            </div>
+                {t("Browse")}
+              </span>
+            </p>
+            <small className={dash_styles.small}>
+              {t("Supported Formats: MP3")}
+            </small>
+            {selectedFile && (
+              <p className={dash_styles.selectedFile}>{selectedFile.name}</p>
+            )}
           </div>
-        </main>
+          <input
+            type="file"
+            ref={fileInputRef}
+            onChange={handleFileInput}
+            accept="audio/*"
+            style={{ display: "none" }}
+          />
+          <button
+            className={dash_styles.uploadButton}
+            onClick={handleUpload}
+            disabled={!selectedFile}
+          >
+            {t("Upload Files")}
+          </button>
+        </div>
       </div>
-    </div>
+    </>
   );
 };
 
-GlobalCore.manager.app('dashboard', Dashboard)
+GlobalCore.manager.app("dashboard", Dashboard);
